+++ CHANGELOG.md
@@ -32,6 +32,9 @@ Full Keycloak upstream jira issue can be shown if filtered by Fix version.
 - Add user session name to ClaimToUserSessionNoteMapper
 - Refresh token flow may check oidc idp  refresh token valid
 - Refresh SAML federation and auto-updated IdP
+- Logo uri for IdPs
+- Add cookie for chosen login IdPs
+- Search without accents in Identity Providers
 
 ### Changed
 - Change emailVerified User field with UserAttributeMappers (conditional trust email). [EOSC-KC-70](https://github.com/eosc-kc/keycloak/issues/70)